Do not use \s or [\s] with grep
grep does not appear to treat `\s` as a shorthand character to
match any whitespace.

The prior test was failing to match conditions that it should have
matched, and the failure was resulting in multiple `sha512` tokens
being placed into `/etc/pam.d/system-auth-ac` on the initial run
of the formula.

Fixes #134