V38513: Manage iptables safety rules in ASH chain
Previously, V38513.sls used a jinja test to check for the existence
of `/etc/sysconfig/iptables`, and only created the safety rules for
iptables if the file was missing. However, the INPUT table was
_always_ being configured to DROP. For systems where this file
already existed, this combination caused the system to become
unreachable because the safety rules were not being added.

This patch updates V38513.sls to use the newer functions and
parameters available in the `iptables` state module to create a new
chain, ASH, jump from the INPUT table to the ASH chain, and creates
the safety rules in the ASH chain.

Creating the saftey rules is conditional on whether the jump rule
is added to the INPUT chain. This means users can modify the ASH
chain after the fact if they wish to change the default safety
rules.

Fixes #144